+++ CHANGELOG.md
@@ -1,1 +1,6 @@
 # Changelog
+
+## [0.0.1] - 2019-05-09
+- **New**
+    - Add COM server initialization
+    - Add COM server initialization tests